requirements.txt
gunicorn==19.9.0
Flask==1.1.1
Flask-WTF==0.14.3
h5py==2.10.0
html5lib==0.9999999
itsdangerous==1.1.0
jedi==0.11.1
Jinja2==2.11.1
joblib==0.13.2
Keras==2.4.3
Keras-Applications==1.0.2
Keras-Preprocessing==1.1.2
numpy==1.18.5

Pillow==7.1.2


tensorflow==2.3.0

Werkzeug==1.0.0
widgetsnbextension==3.2.0

WTForms==2.3.3